+++ src/Stack/SetupCmd.hs
@@ -34,9 +34,22 @@ data SetupCmdOpts = SetupCmdOpts
     { scoCompilerVersion :: !(Maybe CompilerVersion)
     , scoForceReinstall  :: !Bool
     , scoUpgradeCabal    :: !Bool
-    , scoStackSetupYaml  :: !String
+    , scoSetupInfoYaml   :: !String
     , scoGHCBindistURL   :: !(Maybe String)
     }
+
+setupYamlCompatParser :: OA.Parser String
+setupYamlCompatParser = stackSetupYaml <|> setupInfoYaml
+    where stackSetupYaml = OA.strOption (
+               OA.long "stack-setup-yaml"
+            <> OA.help "DEPRECATED: Use 'setup-info-yaml' instead"
+            <> OA.metavar "URL"
+            <> OA.hidden )
+          setupInfoYaml  = OA.strOption (
+               OA.long "setup-info-yaml"
+            <> OA.help "Alternate URL or absolute path for stack dependencies"
+            <> OA.metavar "URL"
+            <> OA.value defaultSetupInfoYaml )
 
 setupParser :: OA.Parser SetupCmdOpts
 setupParser = SetupCmdOpts
@@ -52,11 +65,7 @@ setupParser = SetupCmdOpts
             "upgrade-cabal"
             "installing the newest version of the Cabal library globally"
             OA.idm
-    <*> OA.strOption
-            ( OA.long "stack-setup-yaml"
-           <> OA.help "Location of the main stack-setup.yaml file"
-           <> OA.value defaultStackSetupYaml
-           <> OA.showDefault )
+    <*> setupYamlCompatParser
     <*> OA.optional (OA.strOption
             (OA.long "ghc-bindist"
            <> OA.metavar "URL"
@@ -94,7 +103,7 @@ setup SetupCmdOpts{..} wantedCompiler compilerCheck mstack = do
         , soptsSkipMsys = configSkipMsys
         , soptsUpgradeCabal = scoUpgradeCabal
         , soptsResolveMissingGHC = Nothing
-        , soptsStackSetupYaml = scoStackSetupYaml
+        , soptsSetupInfoYaml = scoSetupInfoYaml
         , soptsGHCBindistURL = scoGHCBindistURL
         }
     let compiler = case wantedCompiler of